Fixed compiler warnung about type norrowing

+++ SU2_CFD/src/output/COutput.cpp
@@ -149,7 +149,7 @@ COutput::COutput(const CConfig *config, unsigned short ndim, bool fem_output):
   wndCauchyEps     = config->GetWnd_Cauchy_Eps();
 
   wndConvFields.reserve(config->GetnWndConv_Field());
-  for (unsigned short iField = 0; iField < config->GetnWndConv_Field(); iField++){
+  for (size_t iField = 0; iField < config->GetnWndConv_Field(); iField++){
     wndConvFields.emplace_back(config->GetWndConv_Field(iField));
   }
